Refactor
Pulled AddPlaceholderForSubFolders function out of the code to add the
drives to the directory tree.

First monolithic function is now broken down into 4 functions, the
longest of which is about 14 lines. That's a hell of a lot more
readable. It's not perfect, but nothing is. I'm just cleaning this up
first, not wanting to re-write it totally.

Honestly wondering how in the merry hell I read this mess when I first
wrote it. Granted, I did this probably 5 years ago, but still. I feel
the need to apologize to anyone who tried to read this code. They
probably wanted to smack me.

Bad llama. Bad bad llama - https://www.youtube.com/watch?v=YQrhXSehUrg