Bugfix
Fixed bug that occurred if you tried to expand a folder that previously
had sub folders, but the sub folders were deleted.

I never considered this edge case because I'm a dummy. Of course, nobody
ever reported it to me either, so either nobody bothered to or I'm the
first person to run into this edge case.

The function that contained this error still needs to be refactored.
Badly.