Refactor
Found duplicate code in the fileTree_BeforeExpand function.

I had already pulled code that was almost exactly the same into the
function AddPlaceholderForSubFolders in a previous refactor.

Changed AddPlaceholderForSubFolders very slightly to allow code reuse
and added a call to it in fileTree_BeforeExpand in place of the
duplicated code.

Hulk take OOP stick and HULK SMASH DUPLICATED CODE. Hulk wonder what the
hell past Hulk was thinking when writing this. Bad hulk.