Refactor
Okay, I lied about being done. There was one more piece of code that was
bugging the crap out of me.

Pulled the code out of fileTree_BeforeExpand that adds the actual sub
directories to the tree when you expand a folder and made it into its
own function.

It increases readability a bit since you don't have to sit there for 10
seconds to figure out what that block of code is doing.